PRSD-623: Split Integration Tests into Journey and Single Page Tests (#439)
