+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/dataModels/AddressDataModel.kt
@@ -59,6 +59,4 @@ data class AddressDataModel(
                 postcode = address.postcode,
             )
     }
-
-    fun getHouseNameOrNumber() = this.buildingName ?: this.buildingNumber
 }
