+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/CancelLaUserInvitationTests.kt
@@ -2,8 +2,9 @@ package uk.gov.communities.prsdb.webapp.integration
 
 import com.microsoft.playwright.Page
 import com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat
+import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
-import org.springframework.test.context.jdbc.Sql
+import uk.gov.communities.prsdb.webapp.integration.JourneyTestWithSeedData.NestedJourneyTestWithSeedData
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.assertThat
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.CancelLaUserInvitationPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.CancelLaUserInvitationSuccessPage
@@ -14,58 +15,68 @@ import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.ManageLaUse
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage.Companion.assertPageIs
 
 class CancelLaUserInvitationTests : IntegrationTest() {
-    @Test
-    @Sql("/data-la-users-and-invitations.sql")
-    fun `an la user invitation can be cancelled`(page: Page) {
-        // Changing the pending user takes you to the cancel invitation page
-        var pendingInvitationRowIndex = 2
-        var manageUsersPage = navigator.goToManageLaUsers(1)
-        assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, ACCOUNT_STATUS_COL_INDEX)).containsText("PENDING")
-        assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, USERNAME_COL_INDEX)).containsText("invited.user@example.com")
-        manageUsersPage.getChangeLink(rowIndex = pendingInvitationRowIndex).clickAndWait()
-        val cancelInvitationPage = assertPageIs(page, CancelLaUserInvitationPage::class)
+    @Nested
+    inner class LaUserInvitation : NestedJourneyTestWithSeedData("data-la-users-and-invitations.sql") {
+        @Test
+        fun `an la user invitation can be cancelled`(page: Page) {
+            // Changing the pending user takes you to the cancel invitation page
+            val pendingInvitationRowIndex = 2
+            var manageUsersPage = navigator.goToManageLaUsers(1)
+            assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, ACCOUNT_STATUS_COL_INDEX)).containsText("PENDING")
+            assertThat(
+                manageUsersPage.table.getCell(pendingInvitationRowIndex, USERNAME_COL_INDEX),
+            ).containsText("invited.user@example.com")
+            manageUsersPage.getChangeLink(rowIndex = pendingInvitationRowIndex).clickAndWait()
+            val cancelInvitationPage = assertPageIs(page, CancelLaUserInvitationPage::class)
 
-        // Cancel invitation
-        assertThat(cancelInvitationPage.userDetailsSection).containsText("invited.user@example.com")
-        cancelInvitationPage.form.submit()
-        val successPage = assertPageIs(page, CancelLaUserInvitationSuccessPage::class)
+            // Cancel invitation
+            assertThat(cancelInvitationPage.userDetailsSection).containsText("invited.user@example.com")
+            cancelInvitationPage.form.submit()
+            val successPage = assertPageIs(page, CancelLaUserInvitationSuccessPage::class)
 
-        // The success page confirms the user is deleted
-        assertThat(successPage.confirmationBanner).containsText("You've cancelled invited.user@example.com's invitation from ISLE OF MAN")
-        successPage.returnButton.clickAndWait()
-        manageUsersPage = assertPageIs(page, ManageLaUsersPage::class)
+            // The success page confirms the user is deleted
+            assertThat(
+                successPage.confirmationBanner,
+            ).containsText("You've cancelled invited.user@example.com's invitation from ISLE OF MAN")
+            successPage.returnButton.clickAndWait()
+            manageUsersPage = assertPageIs(page, ManageLaUsersPage::class)
 
-        // The invited user is no longer in the table
-        assertThat(
-            manageUsersPage.table.getCell(pendingInvitationRowIndex, USERNAME_COL_INDEX),
-        ).not().containsText("invited.user@example.com")
+            // The invited user is no longer in the table
+            assertThat(
+                manageUsersPage.table.getCell(pendingInvitationRowIndex, USERNAME_COL_INDEX),
+            ).not().containsText("invited.user@example.com")
+        }
     }
 
-    @Test
-    @Sql("/data-la-invitations-user-is-system-operator.sql")
-    fun `an la admin invitation can be cancelled by a system operator`(page: Page) {
-        // Changing the pending user takes you to the cancel invitation page
-        var pendingInvitationRowIndex = 3
-        var manageUsersPage = navigator.goToManageLaUsers(1)
-        assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, ACCOUNT_STATUS_COL_INDEX)).containsText("PENDING")
-        assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, USERNAME_COL_INDEX)).containsText("x.adminuser@example.com")
-        assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, ACCESS_LEVEL_COL_INDEX)).containsText("Admin")
-        manageUsersPage.getChangeLink(rowIndex = pendingInvitationRowIndex).clickAndWait()
-        val cancelInvitationPage = assertPageIs(page, CancelLaUserInvitationPage::class)
+    @Nested
+    inner class LaAdminInvitation : NestedJourneyTestWithSeedData("data-la-invitations-user-is-system-operator.sql") {
+        @Test
+        fun `an la admin invitation can be cancelled by a system operator`(page: Page) {
+            // Changing the pending user takes you to the cancel invitation page
+            val pendingInvitationRowIndex = 3
+            var manageUsersPage = navigator.goToManageLaUsers(1)
+            assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, ACCOUNT_STATUS_COL_INDEX)).containsText("PENDING")
+            assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, USERNAME_COL_INDEX)).containsText("x.adminuser@example.com")
+            assertThat(manageUsersPage.table.getCell(pendingInvitationRowIndex, ACCESS_LEVEL_COL_INDEX)).containsText("Admin")
+            manageUsersPage.getChangeLink(rowIndex = pendingInvitationRowIndex).clickAndWait()
+            val cancelInvitationPage = assertPageIs(page, CancelLaUserInvitationPage::class)
 
-        // Cancel invitation
-        assertThat(cancelInvitationPage.userDetailsSection).containsText("x.adminuser@example.com")
-        cancelInvitationPage.form.submit()
-        val successPage = assertPageIs(page, CancelLaUserInvitationSuccessPage::class)
+            // Cancel invitation
+            assertThat(cancelInvitationPage.userDetailsSection).containsText("x.adminuser@example.com")
+            cancelInvitationPage.form.submit()
+            val successPage = assertPageIs(page, CancelLaUserInvitationSuccessPage::class)
 
-        // The success page confirms the user is deleted
-        assertThat(successPage.confirmationBanner).containsText("You've cancelled x.adminuser@example.com's invitation from ISLE OF MAN")
-        successPage.returnButton.clickAndWait()
-        manageUsersPage = assertPageIs(page, ManageLaUsersPage::class)
+            // The success page confirms the user is deleted
+            assertThat(
+                successPage.confirmationBanner,
+            ).containsText("You've cancelled x.adminuser@example.com's invitation from ISLE OF MAN")
+            successPage.returnButton.clickAndWait()
+            manageUsersPage = assertPageIs(page, ManageLaUsersPage::class)
 
-        // The invited user is no longer in the table
-        assertThat(
-            manageUsersPage.table.getCell(pendingInvitationRowIndex, USERNAME_COL_INDEX),
-        ).not().containsText("x.adminuser@example.com")
+            // The invited user is no longer in the table
+            assertThat(
+                manageUsersPage.table.getCell(pendingInvitationRowIndex, USERNAME_COL_INDEX),
+            ).not().containsText("x.adminuser@example.com")
+        }
     }
 }
