+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/ErrorPageTests.kt
@@ -3,13 +3,11 @@ package uk.gov.communities.prsdb.webapp.integration
 import com.microsoft.playwright.Page
 import com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat
 import org.junit.jupiter.api.Test
-import org.springframework.test.context.jdbc.Sql
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.ErrorPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage.Companion.createValidPage
 
-@Sql("/data-local.sql")
-class ErrorPageTests : IntegrationTest() {
+class ErrorPageTests : SinglePageTestWithSeedData("data-local.sql") {
     @Test
     fun `500 page renders when error controller path called`(page: Page) {
         navigator.navigate("/error")
