+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/LandlordDashboardTests.kt
@@ -2,7 +2,7 @@ package uk.gov.communities.prsdb.webapp.integration
 
 import com.microsoft.playwright.Page
 import org.junit.jupiter.api.Assertions.assertTrue
-import org.springframework.test.context.jdbc.Sql
+import org.junit.jupiter.api.Nested
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.assertThat
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LandlordDetailsPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LandlordIncompletePropertiesPage
@@ -11,8 +11,7 @@ import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyReg
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-@Sql("/data-local.sql")
-class LandlordDashboardTests : IntegrationTest() {
+class LandlordDashboardTests : SinglePageTestWithSeedData("data-local.sql") {
     @Test
     fun `the dashboard loads displaying the user's name and lrn`() {
         val dashboard = navigator.goToLandlordDashboard()
@@ -57,34 +56,41 @@ class LandlordDashboardTests : IntegrationTest() {
         assertTrue(page.url().contains("https://www.gov.uk/government/publications/guide-to-the-renters-rights-bill"))
     }
 
-    @Test
-    @Sql("/data-mockuser-landlord-with-one-incomplete-property.sql")
-    fun `the dashboard loads with a notification banner and correct message when the landlord has one incomplete property`(page: Page) {
-        val dashboard = navigator.goToLandlordDashboard()
-        assertThat(dashboard.notificationBanner.title).containsText("Important")
-        assertThat(dashboard.notificationBanner.subheading).containsText("You have 1 incomplete property: View incomplete properties")
+    @Nested
+    inner class LandlordWithOneIncompleteProperty :
+        NestedSinglePageTestWithSeedData("data-mockuser-landlord-with-one-incomplete-property.sql") {
+        @Test
+        fun `the dashboard loads with a notification banner and correct message when the landlord has one incomplete property`(page: Page) {
+            val dashboard = navigator.goToLandlordDashboard()
+            assertThat(dashboard.notificationBanner.title).containsText("Important")
+            assertThat(dashboard.notificationBanner.subheading).containsText("You have 1 incomplete property: View incomplete properties")
+        }
     }
 
-    @Test
-    @Sql("/data-mockuser-landlord-with-incomplete-properties.sql")
-    fun `the dashboard loads with a notification banner when the landlord has incomplete properties`(page: Page) {
-        val dashboard = navigator.goToLandlordDashboard()
-        assertThat(dashboard.notificationBanner.title).containsText("Important")
-        assertThat(dashboard.notificationBanner.subheading).containsText("You have 2 incomplete properties: View incomplete properties")
+    @Nested
+    inner class LandlordWithIncompleteProperties :
+        NestedSinglePageTestWithSeedData("data-mockuser-landlord-with-incomplete-properties.sql") {
+        @Test
+        fun `the dashboard loads with a notification banner when the landlord has incomplete properties`(page: Page) {
+            val dashboard = navigator.goToLandlordDashboard()
+            assertThat(dashboard.notificationBanner.title).containsText("Important")
+            assertThat(dashboard.notificationBanner.subheading).containsText("You have 2 incomplete properties: View incomplete properties")
+        }
+
+        @Test
+        fun `the link in the notification banner redirects to the incomplete properties page`(page: Page) {
+            val dashboard = navigator.goToLandlordDashboard()
+            dashboard.notificationBanner.link.clickAndWait()
+            assertPageIs(page, LandlordIncompletePropertiesPage::class)
+        }
     }
 
-    @Test
-    @Sql("/data-mockuser-landlord-with-incomplete-properties.sql")
-    fun `the link in the notification banner redirects to the incomplete properties page`(page: Page) {
-        val dashboard = navigator.goToLandlordDashboard()
-        dashboard.notificationBanner.link.clickAndWait()
-        assertPageIs(page, LandlordIncompletePropertiesPage::class)
-    }
-
-    @Test
-    @Sql("/data-mockuser-landlord-with-properties.sql")
-    fun `the dashboard loads without a notification banner when the landlord has no incomplete properties`() {
-        val dashboard = navigator.goToLandlordDashboard()
-        assertThat(dashboard.notificationBanner).isHidden()
+    @Nested
+    inner class LandlordWithoutIncompleteProperties : NestedSinglePageTestWithSeedData("data-mockuser-landlord-with-properties.sql") {
+        @Test
+        fun `the dashboard loads without a notification banner when the landlord has no incomplete properties`() {
+            val dashboard = navigator.goToLandlordDashboard()
+            assertThat(dashboard.notificationBanner).isHidden()
+        }
     }
 }
