+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/LandlordDetailsUpdateJourneyTests.kt
@@ -11,9 +11,7 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.any
 import org.mockito.kotlin.whenever
-import org.springframework.test.context.jdbc.Sql
 import uk.gov.communities.prsdb.webapp.constants.MANUAL_ADDRESS_CHOSEN
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.assertThat
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LandlordDetailsPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LookupAddressFormPageUpdateLandlordDetails
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.ManualAddressFormPageUpdateLandlordDetails
@@ -22,13 +20,11 @@ import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.B
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.updateLandlordDetailsPages.DateOfBirthFormPageUpdateLandlordDetails
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.updateLandlordDetailsPages.EmailFormPageUpdateLandlordDetails
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.updateLandlordDetailsPages.NameFormPageUpdateLandlordDetails
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.updateLandlordDetailsPages.NoAddressFoundFormPageUpdateLandlordDetails
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.updateLandlordDetailsPages.PhoneNumberFormPageUpdateLandlordDetails
 import uk.gov.communities.prsdb.webapp.models.dataModels.AddressDataModel
 import uk.gov.communities.prsdb.webapp.testHelpers.extensions.getFormattedUkPhoneNumber
 
-@Sql("/data-local.sql")
-class LandlordDetailsUpdateJourneyTests : IntegrationTest() {
+class LandlordDetailsUpdateJourneyTests : JourneyTestWithSeedData("data-local.sql") {
     private val phoneNumberUtil = PhoneNumberUtil.getInstance()
     val addressFound = "Entirely new test address"
 
@@ -50,8 +46,7 @@ class LandlordDetailsUpdateJourneyTests : IntegrationTest() {
     }
 
     @Nested
-    inner class NameUpdates {
-        @Sql("/data-unverified-landlord.sql")
+    inner class NameUpdates : NestedJourneyTestWithSeedData("data-unverified-landlord.sql") {
         @Test
         fun `An unverified landlord can update their name`(page: Page) {
             // Details page
@@ -68,22 +63,10 @@ class LandlordDetailsUpdateJourneyTests : IntegrationTest() {
             // Check changes have occurred
             assertThat(landlordDetailsPage.personalDetailsSummaryList.nameRow.value).containsText(newName)
         }
-
-        @Test
-        fun `A verified landlord cannot update their name`(page: Page) {
-            // Check change link is hidden on details page
-            val landlordDetailsPage = navigator.goToLandlordDetails()
-            assertThat(landlordDetailsPage.personalDetailsSummaryList.nameRow.actions.actionLink).isHidden()
-
-            // Check update name page can't be reached
-            navigator.navigateToLandlordDetailsUpdateNamePage()
-            assertPageIs(page, LandlordDetailsPage::class)
-        }
     }
 
     @Nested
-    inner class DateOfBirthUpdates {
-        @Sql("/data-unverified-landlord.sql")
+    inner class DateOfBirthUpdates : NestedJourneyTestWithSeedData("data-unverified-landlord.sql") {
         @Test
         fun `An unverified landlord can update their date of birth`(page: Page) {
             // Details page
@@ -100,17 +83,6 @@ class LandlordDetailsUpdateJourneyTests : IntegrationTest() {
             // Check changes have occurred
             assertThat(landlordDetailsPage.personalDetailsSummaryList.dateOfBirthRow.value)
                 .containsText(formatDateOfBirth(newDateOfBirth), LocatorAssertions.ContainsTextOptions().setIgnoreCase(true))
-        }
-
-        @Test
-        fun `A verified landlord cannot update their date of birth`(page: Page) {
-            // Check change link is hidden on details page
-            val landlordDetailsPage = navigator.goToLandlordDetails()
-            assertThat(landlordDetailsPage.personalDetailsSummaryList.dateOfBirthRow.actions.actionLink).isHidden()
-
-            // Check update date of birth page can't be reached
-            navigator.navigateToLandlordDetailsUpdateDateOfBirthPage()
-            assertPageIs(page, LandlordDetailsPage::class)
         }
     }
 
@@ -205,40 +177,6 @@ class LandlordDetailsUpdateJourneyTests : IntegrationTest() {
             // Check changes have occurred
             val newSingleLineAddress = AddressDataModel.manualAddressDataToSingleLineAddress(newFirstLine, newTown, newPostcode)
             assertThat(landlordDetailsPage.personalDetailsSummaryList.addressRow.value).containsText(newSingleLineAddress)
-        }
-
-        @Test
-        fun `A landlord can search again via the Select Address page`(page: Page) {
-            val selectAddressPage = navigator.skipToLandlordDetailsUpdateSelectAddressPage()
-            selectAddressPage.searchAgain.clickAndWait()
-            assertPageIs(page, LookupAddressFormPageUpdateLandlordDetails::class)
-        }
-
-        @Test
-        fun `A landlord can search again or choose manual address via the No Address Found page if no addresses are found`(page: Page) {
-            // Arrange for no addresses to be found
-            val houseNumber = "15"
-            val postcode = "AB1 2CD"
-            whenever(osPlacesClient.search(houseNumber, postcode)).thenReturn("{}")
-
-            // Lookup Address page
-            val lookupAddressPage = navigator.goToUpdateLandlordDetailsUpdateLookupAddressPage()
-            lookupAddressPage.submitPostcodeAndBuildingNameOrNumber(postcode, houseNumber)
-            var noAddressFoundPage = assertPageIs(page, NoAddressFoundFormPageUpdateLandlordDetails::class)
-
-            // No Address Found page
-            assertThat(noAddressFoundPage.heading).containsText(houseNumber)
-            assertThat(noAddressFoundPage.heading).containsText(postcode)
-
-            // Search again
-            noAddressFoundPage.searchAgain.clickAndWait()
-            val lookupAddressPageAgain = assertPageIs(page, LookupAddressFormPageUpdateLandlordDetails::class)
-            lookupAddressPageAgain.submitPostcodeAndBuildingNameOrNumber(postcode, houseNumber)
-            noAddressFoundPage = assertPageIs(page, NoAddressFoundFormPageUpdateLandlordDetails::class)
-
-            // Choose Manual Address
-            noAddressFoundPage.form.submit()
-            assertPageIs(page, ManualAddressFormPageUpdateLandlordDetails::class)
         }
     }
