+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/LandlordIncompletePropertiesPageTests.kt
@@ -5,8 +5,8 @@ import com.microsoft.playwright.assertions.LocatorAssertions
 import kotlinx.datetime.DatePeriod
 import kotlinx.datetime.plus
 import org.junit.jupiter.api.Nested
-import org.springframework.test.context.jdbc.Sql
 import uk.gov.communities.prsdb.webapp.helpers.DateTimeHelper
+import uk.gov.communities.prsdb.webapp.integration.SinglePageTestWithSeedData.NestedSinglePageTestWithSeedData
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.assertThat
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.DeleteIncompletePropertyRegistrationAreYouSurePage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LandlordDetailsPage
@@ -18,8 +18,8 @@ import kotlin.test.assertEquals
 
 class LandlordIncompletePropertiesPageTests : IntegrationTest() {
     @Nested
-    @Sql("/data-mockuser-landlord-with-incomplete-properties.sql")
-    inner class LandlordsWithIncompleteProperties {
+    inner class LandlordsWithIncompleteProperties :
+        NestedSinglePageTestWithSeedData("data-mockuser-landlord-with-incomplete-properties.sql") {
         @Test
         fun `the page loads with heading and subheading`() {
             val incompletePropertiesPage = navigator.goToLandlordIncompleteProperties()
@@ -73,8 +73,7 @@ class LandlordIncompletePropertiesPageTests : IntegrationTest() {
     }
 
     @Nested
-    @Sql("/data-mockuser-landlord-with-properties.sql")
-    inner class LandlordsWithNoIncompleteProperties {
+    inner class LandlordsWithNoIncompleteProperties : NestedSinglePageTestWithSeedData("data-mockuser-landlord-with-properties.sql") {
         @Test
         fun `the page loads with heading and page text`() {
             val incompletePropertiesPage = navigator.goToLandlordIncompleteProperties()
