+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/RegisterHomePageTests.kt
@@ -3,10 +3,8 @@ package uk.gov.communities.prsdb.webapp.integration
 import com.microsoft.playwright.Page
 import com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat
 import org.junit.jupiter.api.Test
-import org.springframework.test.context.jdbc.Sql
 
-@Sql("/data-local.sql")
-class RegisterHomePageTests : IntegrationTest() {
+class RegisterHomePageTests : SinglePageTestWithSeedData("data-local.sql") {
     @Test
     fun `register a home to rent page renders`(page: Page) {
         page.navigate("http://localhost:$port/registration")
